# src/main/resources/openapi/weather-app.yaml
openapi: 3.0.2
info:
  title: Weather application
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Development server
security:
  - cookieAuth: []
paths:
...
  /region/types/{typeId}:
    get:
      summary: Получение информации о типе региона
      tags:
        - RegionType
      operationId: getRegionType
      parameters:
        - name: typeId
          in: path
          description: Идентификатор типа региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionTypeResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тип региона с таким typeId не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Изменение типа региона
      tags:
        - RegionType
      operationId: updateRegionType
      parameters:
        - name: typeId
          in: path
          description: Идентификатор типа региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionTypeRequest'
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionTypeResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тип региона с таким typeId не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Тип региона с таким type уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Удаление типа региона
      tags:
        - RegionType
      operationId: deleteRegionType
      parameters:
        - name: typeId
          in: path
          description: Идентификатор типа региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionTypeResponse'
        '400':
          description: Ошибка валидации | Есть регионы с типом с typeId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Тип региона с таким typeId не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /region/types:
    post:
      summary: Добавление типа региона
      tags:
        - RegionType
      operationId: createRegionType
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionTypeRequest'
      responses:
        '201':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionTypeResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Тип региона с таким type уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /region/weather/{regionId}:
    get:
      summary: Получение информации о погоде в регионе
      tags:
        - Weather
      operationId: getRegionWeather
      parameters:
        - name: regionId
          in: path
          description: Идентификатор региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionWeatherResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Регион с regionld не найден или forecastld в данном регионе не найденна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      summary: Изменение погоды
      tags:
        - Weather
      operationId: updateRegionWeather
      parameters:
        - name: regionId
          in: path
          description: Идентификатор региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateRegionWeatherRequest'
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionWeatherResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Регион с regionld не найден или forecastld в данном регионе не найденна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Удаление погоды
      tags:
        - Weather
      operationId: deleteRegionWeather
      parameters:
        - name: regionId
          in: path
          description: Идентификатор региона
          required: true
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 1
      responses:
        '200':
          description: Запрос успешно выполнен
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Регион с regionld не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /region/weather:
    post:
      summary: Добавление погоды
      tags:
        - Weather
      operationId: addWeather
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegionWeatherRequest'
      responses:
        '200':
          description: Запрос успешно выполнен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostRegionWeatherResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Неверные авторизационные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Регион с regionld не найден или forecastld в данном регионе не найденна
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: id
  schemas:
    ErrorResponse:
      type: object
      required:
        - status
        - message
        - timestamp
        - path
      properties:
        status:
          type: integer
          example: 404
          description: Http status code
        message:
          type: string
          description: Error message
        timestamp:
          type: string
          format: date-time
          description: Error occurrence time
        path:
          type: string
          description: URI path
        details:
          type: object
          description: Error details
          additionalProperties:
            type: string
...
    RegionTypeRequest:
      type: object
      required:
        - type
      properties:
        type:
          type: string
          description: Тип региона
    RegionTypeResponse:
      type: object
      required:
        - id
        - type
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор типа региона
        type:
          type: string
          description: Тип региона
    WeatherCondition:
      type: string
      description: Текущее состояние погоды
      enum:
        - CLEAR
        - CLOUDY
        - RAIN
        - SNOW
        - FOG
        - STORM
...
    RegionWeatherRequest:
      type: object
      required:
        - regionId
        - temperature
        - humidity
        - windSpeed
        - weatherCondition
        - precipitationAmount
        - measurementDateTime
      properties:
        regionId:
          type: integer
          format: int64
          minimum: 1
          description: Идентификатор региона
        temperature:
          type: number
          format: float
          description: Температура в регионе, °C
        humidity:
          type: number
          format: float
          description: Влажность воздуха в регионе, %
        windSpeed:
          type: number
          format: float
          minimum: 0
          description: Скорость ветра, м/с
        weatherCondition:
          $ref: '#/components/schemas/WeatherCondition'
        precipitationAmount:
          type: number
          format: float
          minimum: 0
          description: Количество осадков, мм
        measurementDateTime:
          type: string
          format: date-time
          description: Дата и время измерения погодных условий в формате ISO-8601
        weatherForecast:
          type: array
          description: Массив идентификаторов объектов с прогнозом погоды на ближайшие дни
          items:
            type: integer
            format: int64
    RegionWeatherResponse:
      type: object
      required:
        - id
        - regionName
        - temperature
        - humidity
        - windSpeed
        - weatherCondition
        - precipitationAmount
        - measurementDateTime
      properties:
        id:
          type: integer
          format: int64
          description: Идентификатор региона
        regionName:
          type: string
          description: Название региона
        temperature:
          type: number
          format: float
          description: Температура в регионе, °C
        humidity:
          type: number
          format: float
          description: Влажность воздуха в регионе, %
        windSpeed:
          type: number
          format: float
          description: Скорость ветра, м/с
        weatherCondition:
          $ref: '#/components/schemas/WeatherCondition'
        precipitationAmount:
          type: number
          format: float
          description: Количество осадков, мм
        measurementDateTime:
          type: string
          format: date-time
          description: Дата и время измерения погодных условий в формате ISO-8601
        weatherForecast:
          type: array
          description: Массив идентификаторов объектов с прогнозом погоды на ближайшие дни
          items:
            type: integer
            format: int64
    PostRegionWeatherResponse:
      type: object
      required:
        - id
        - temperature
        - humidity
        - windSpeed
        - weatherCondition
        - precipitationAmount
        - measurementDateTime
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор созданной записи о погоде в регионе
        temperature:
          type: number
          format: float
          description: Температура в регионе, °C
        humidity:
          type: number
          format: float
          description: Влажность воздуха в регионе, %
        windSpeed:
          type: number
          format: float
          description: Скорость ветра, м/с
        weatherCondition:
          $ref: '#/components/schemas/WeatherCondition'
        precipitationAmount:
          type: number
          format: float
          description: Количество осадков, мм
        measurementDateTime:
          type: string
          format: date-time
          description: Дата и время измерения погодных условий в формате ISO-8601
        weatherForecast:
          type: array
          description: Массив идентификаторов объектов с прогнозом погоды на ближайшие дни
          items:
            type: integer
            format: int64
    UpdateRegionWeatherRequest:
      type: object
      required:
        - temperature
        - humidity
        - windSpeed
        - weatherCondition
        - precipitationAmount
        - measurementDateTime
      properties:
        temperature:
          type: number
          format: float
          description: Температура в регионе, °C
        humidity:
          type: number
          format: float
          description: Влажность воздуха в регионе, %
        windSpeed:
          type: number
          format: float
          minimum: 0
          description: Скорость ветра, м/с
        weatherCondition:
          $ref: '#/components/schemas/WeatherCondition'
        precipitationAmount:
          type: number
          format: float
          minimum: 0
          description: Количество осадков, мм
        measurementDateTime:
          type: string
          format: date-time
          description: Дата и время измерения погодных условий в формате ISO-8601
        weatherForecast:
          type: array
          description: Массив идентификаторов объектов с прогнозом погоды на ближайшие дни
          items:
            type: integer
            format: int64
...